// src/shared/types.ts
export interface StickerItem {
  id: string;
  accessHash: string;
  alt: string;
  url: string;
  type: 'static' | 'animated' | 'video';
  mimeType: string;
  width?: number;
  height?: number;
}

export interface StickerSet {
  id: string;
  title: string;
  shortName: string;
  count: number;
  isAnimated?: boolean;
  isVideo?: boolean;
  thumbnailUrl?: string;
  stickers: StickerItem[];
}